requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
asyncpg>=0.30.0
python-dotenv>=1.0.0
redis[hiredis]>=5.0.0
polars>=1.18.0
orjson>=3.10.0
